feat: sort tasks by status (open first) + move landing to root
- Tasks page now sorts by status: open first, then closed
- Moved /landing content to / (root)
- Removed redundant landing folder
